feat(skill): 优化 Skill 工具可见性和上下文持久化
